Data pipeline/Cypher Generation/requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pathlib2>=2.3.7
